replace the callable `parse_bytes` parameter with `type_code`
